+++ backend/labelling_application/src/main/java/com/lbls/labelling_application/service/mer_data/MerDataService.java
@@ -6,5 +6,5 @@ public interface MerDataService {
 
     MerData save(MerData merData);
 
-    MerData findById(Long id);
+    MerData findById(int id);
 }
